Add functionality to connect with multiple databases concurrently

+++ src/connectionpool.cpp
@@ -3,6 +3,7 @@
 
 #include "dbexception.h"
 #include "config/configbuilder.h"
+#include "config/configmanager.h"
 
 #include <QtSql/QSqlQuery>
 #include <QtSql/QSqlError>
@@ -14,32 +15,56 @@ QTDAO_USING_NAMESPACE
  * @param name current thread connection name
  */
 ConnectionData::ConnectionData(QString name)
-    : connectionName(std::move(name))
-{
-    openNewConnection();
-}
+    : baseConnectionName(std::move(name))
+{}
 
 /**
  * remove current thread used connection, and push into used names
  */
 ConnectionData::~ConnectionData() {
-    ConnectionPool::closeConnection(connectionName);
+    ConnectionPool::closeConnection(baseConnectionName, usedConnectionNames.values());
 }
 
 /**
  * get saved connection, and test connection useful
+ * @param sessionId session for different database connection in current thread
  * @return opened connection
  */
-QSqlDatabase ConnectionData::getOpenedConnection() {
+QSqlDatabase ConnectionData::getOpenedConnection(qint64 sessionId) {
+    if (sessionId == -1) {
+        if (!holdSessionIdStack.isEmpty()) {
+            currentProcessSessionId = holdSessionIdStack.top();
+        }
+    }
+    currentProcessSessionId = sessionId;
     testConnection();
-    return QSqlDatabase::database(connectionName);
+    return QSqlDatabase::database(usedConnectionNames.value(sessionId));
+}
+
+/**
+ * push new session id use by scope connection
+ * @param sessionId
+ */
+void ConnectionData::beginSession(qint64 sessionId) {
+    holdSessionIdStack.push(sessionId);
+}
+
+/**
+ * pop session id use by scope connection
+ */
+void ConnectionData::endSession() {
+    if (!holdSessionIdStack.isEmpty()) {
+        holdSessionIdStack.pop();
+    }
 }
 
 /**
  * create new database connect use current bind connection name
  */
 void ConnectionData::openNewConnection() {
-    ConnectionPool::createConnection(connectionName);
+    auto connectionName = nameForDatabase();
+    ConnectionPool::createConnection(currentProcessSessionId, connectionName);
+    usedConnectionNames.insert(currentProcessSessionId, connectionName);
 }
 
 /**
@@ -47,17 +72,29 @@ void ConnectionData::openNewConnection() {
  */
 void ConnectionData::testConnection() {
     //test connection useful
-    if (QSqlDatabase::contains(connectionName)) {
-        {
-            auto db = QSqlDatabase::database(connectionName);
-            QSqlQuery query("select 1", db);
-            if (query.lastError().type() == QSqlError::NoError) {
-                return;
+    if (usedConnectionNames.contains(currentProcessSessionId)) {
+        const auto& connectionName = usedConnectionNames.value(currentProcessSessionId);
+        if (QSqlDatabase::contains(connectionName)) {
+            {
+                auto db = QSqlDatabase::database(connectionName);
+                QSqlQuery query("select 1", db);
+                if (query.lastError().type() == QSqlError::NoError) {
+                    return;
+                }
             }
+            QSqlDatabase::removeDatabase(connectionName);
         }
-        QSqlDatabase::removeDatabase(connectionName);
+        usedConnectionNames.remove(currentProcessSessionId);
     }
     openNewConnection();
+}
+
+/**
+ * create connection name used by session id
+ * @return
+ */
+QString ConnectionData::nameForDatabase() const {
+    return baseConnectionName + "#" + QString::number(currentProcessSessionId);
 }
 
 QThreadStorage<ConnectionData*> ConnectionPool::localConnection;
@@ -76,36 +113,27 @@ ConnectionPool& ConnectionPool::getInstance() {
  * get current thread saved database connection
  * @return opened connection
  */
-QSqlDatabase ConnectionPool::getConnection() {
+QSqlDatabase ConnectionPool::getConnection(qint64 sessionId) {
+    initializeLocalData();
+    return localConnection.localData()->getOpenedConnection(sessionId);
+}
 
-    if (!localConnection.hasLocalData()) {
-        QString connectionName;
+/**
+ * push new session id in stack
+ * @param sessionId
+ */
+void ConnectionPool::scopeSessionBegin(qint64 sessionId) {
+    Q_ASSERT(sessionId != -1);
+    initializeLocalData();
+    localConnection.localData()->beginSession(sessionId);
+}
 
-        { //locked only read connection name
-            auto &pool = ConnectionPool::getInstance();
-            QMutexLocker locker(&pool.mutex);
-            if (!pool.unusedConnectionNames.empty()) {
-                //use unused connection
-                connectionName = pool.unusedConnectionNames.dequeue();
-                MESSAGE_DEBUG("use recycled connection:", connectionName);
-            } else {
-                //create new connection name
-                connectionName = QString("Connection-%1").arg(++pool.usedConnectionSize);
-                MESSAGE_DEBUG("use new connection:", connectionName);
-            }
-        }
-
-        ConnectionData* data;
-        try {
-            data = new ConnectionData(connectionName);
-            localConnection.setLocalData(data);
-        } catch (DaoException& e) {
-            delete data; //will recycle this connection name
-            throw e;
-        }
-    }
-
-    return localConnection.localData()->getOpenedConnection();
+/**
+ * pop session id in stack
+ */
+void ConnectionPool::scopeSessionEnd() {
+    initializeLocalData();
+    localConnection.localData()->endSession();
 }
 
 /**
@@ -123,6 +151,22 @@ void ConnectionPool::closeAllConnection() {
     pool.unusedConnectionNames.clear();
 }
 
+/**
+ * close all connections reference to the session id
+ * @param sessionId
+ */
+void ConnectionPool::closeConnection(qint64 sessionId) {
+    auto connectNameSuffix = "#" + QString::number(sessionId);
+
+    auto connections = QSqlDatabase::connectionNames();
+    for (const auto& name : connections) {
+        if (name.endsWith(connectNameSuffix)) {
+            QSqlDatabase::removeDatabase(name);
+        }
+    }
+    //It is not necessary to recycle the connection name here
+}
+
 int ConnectionPool::getUsedConnectionSize() {
     auto& pool = ConnectionPool::getInstance();
     QMutexLocker locker(&pool.mutex);
@@ -130,16 +174,43 @@ int ConnectionPool::getUsedConnectionSize() {
 }
 
 /**
- * close connect of current thread, and push into unused names
- * @param connectionName
+ * check current thread data is not empty, or create new connection data use recycled connection
  */
-void ConnectionPool::closeConnection(const QString& connectionName) {
+void ConnectionPool::initializeLocalData() {
+    if (!localConnection.hasLocalData()) {
+        QString connectionName;
+
+        { //locked only read connection name
+            auto &pool = ConnectionPool::getInstance();
+            QMutexLocker locker(&pool.mutex);
+            if (!pool.unusedConnectionNames.empty()) {
+                //use unused connection
+                connectionName = pool.unusedConnectionNames.dequeue();
+                MESSAGE_DEBUG("use recycled connection:", connectionName);
+            } else {
+                //create new connection name
+                connectionName = QString("Connection-%1").arg(++pool.usedConnectionSize);
+                MESSAGE_DEBUG("use new connection:", connectionName);
+            }
+        }
+        localConnection.setLocalData(new ConnectionData(connectionName));
+    }
+}
+
+/**
+ * close connect of current thread, and push into unused names
+ * @param connectionName name used by pool
+ * @param nameInDatabase name used by database
+ */
+void ConnectionPool::closeConnection(const QString& connectionName, const QStringList& nameInDatabase) {
     auto& pool = ConnectionPool::getInstance();
     QMutexLocker locker(&pool.mutex);
 
     pool.unusedConnectionNames.enqueue(connectionName);
-    if (QSqlDatabase::contains(connectionName)) {
-        QSqlDatabase::removeDatabase(connectionName);
+    for (const auto& name : nameInDatabase) {
+        if (QSqlDatabase::contains(name)) {
+            QSqlDatabase::removeDatabase(name);
+        }
     }
 
     MESSAGE_DEBUG("recycle connection:", connectionName);
@@ -150,7 +221,7 @@ void ConnectionPool::closeConnection(const QString& connectionName) {
  * @param connectionName
  * @return opened connection
  */
-QSqlDatabase ConnectionPool::createConnection(const QString &connectionName) {
+QSqlDatabase ConnectionPool::createConnection(qint64 sessionId, const QString &connectionName) {
     // check connection name in QSqlDatabase
     QSqlError lastErr;
     int testCount = 3;
@@ -161,7 +232,7 @@ QSqlDatabase ConnectionPool::createConnection(const QString &connectionName) {
                 db = QSqlDatabase::database(connectionName);
             } else {
                 // create new connection
-                db = prepareConnect(connectionName, globalConfig->mDatabaseName);
+                db = prepareConnect(connectionName, sessionId);
             }
             if (db.open()) {
                 //test opened connection
@@ -183,7 +254,7 @@ QSqlDatabase ConnectionPool::createConnection(const QString &connectionName) {
     throw DaoException(lastErr);
 }
 
-QSqlDatabase ConnectionPool::prepareConnect(const QString& connectName, const QString& dbName) {
-    return globalConfig->getConnection(connectName, dbName);
+QSqlDatabase ConnectionPool::prepareConnect(const QString& connectName, qint64 sessionId) {
+    return ConfigManager::getConfig(sessionId)->getConnection(connectName);
 }
